add complaints module

+++ src/main/java/tn/enicarthage/eniconnect_backend/services/impl/QuestionServiceImpl.java
@@ -1,10 +1,11 @@
 package tn.enicarthage.eniconnect_backend.services.impl;
 
-import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import tn.enicarthage.eniconnect_backend.dtos.request.question.QuestionTemplateDto;
 import tn.enicarthage.eniconnect_backend.entities.QuestionTemplate;
+import tn.enicarthage.eniconnect_backend.exceptions.InvalidDataException;
+import tn.enicarthage.eniconnect_backend.exceptions.ResourceNotFoundException;
 import tn.enicarthage.eniconnect_backend.mappers.QuestionMapper;
 import tn.enicarthage.eniconnect_backend.repositories.QuestionTemplateRepository;
 import tn.enicarthage.eniconnect_backend.services.QuestionService;
@@ -27,7 +28,7 @@ public class QuestionServiceImpl implements QuestionService {
 
     @Override
     public QuestionTemplate createQuestion(QuestionTemplateDto question) {
-        int  order = (int) questionTemplateRepository.count() ;
+        int order = (int) questionTemplateRepository.count();
 
         QuestionTemplate questionTemplate = questionMapper.toEntity(question);
         questionTemplate.setOrderIndex(order);
@@ -35,19 +36,25 @@ public class QuestionServiceImpl implements QuestionService {
     }
 
     @Override
-    public QuestionTemplate updateeQuestionById(long id , QuestionTemplateDto question) {
-        QuestionTemplate questionTemplate = questionTemplateRepository.findById(id).orElseThrow(()-> new RuntimeException("Question Not Found"));
-        questionTemplate.setText(question.text());
-        questionTemplate.setMaxValue(question.maxValue());
-        questionTemplate.setMinValue(question.minValue());
+    public QuestionTemplate updateeQuestionById(long id, QuestionTemplateDto question) {
+        QuestionTemplate questionTemplate = questionTemplateRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Question", "id", id));
 
-        QuestionTemplate updatedQuestion =  questionTemplateRepository.save(questionTemplate);
-        return updatedQuestion ;
+        try {
+            questionTemplate.setText(question.text());
+            questionTemplate.setMaxValue(question.maxValue());
+            questionTemplate.setMinValue(question.minValue());
+            return questionTemplateRepository.save(questionTemplate);
+        } catch (Exception e) {
+            throw new InvalidDataException("Invalid question data: " + e.getMessage());
+        }
     }
 
     @Override
     public void DeleteQuestion(Long id) {
-        QuestionTemplate question = questionTemplateRepository.findById(id).orElseThrow(()->new RuntimeException("Question not found"));
+        if (!questionTemplateRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Question", "id", id);
+        }
         questionTemplateRepository.deleteById(id);
     }
 }
